E_learning/requirements.txt
bcrypt==3.2.2
Bcrypt-Flask==1.0.2
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cx-Logging==3.0
darkdetect==0.5.1
decorator==5.1.1
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
Flask-Admin==1.6.0
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
future==0.18.2
geocoder==1.38.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
lief==0.12.1
MarkupSafe==2.1.1
Pillow==9.1.0
Pillow-PIL==0.1.dev0
ratelim==0.1.6
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.36
urllib3==1.26.9
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
